requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.2
anyio==4.6.2.post1
appdirs==1.4.4
astroid==3.3.5
astunparse==1.6.3
asyncssh==2.17.0
atpublic==5.0
attrs==24.2.0
backoff==2.2.1
billiard==4.2.1
blinker==1.8.2
boto3==1.35.42
botocore==1.35.42
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.0
contractions==0.1.73
cryptography==43.0.1
cycler==0.12.1
dacite==1.6.0
dagshub==0.3.39
dagshub-annotation-converter==0.1.1
databricks-sdk==0.34.0
dataclasses-json==0.6.7
Deprecated==1.2.14
dictdiffer==0.9.0
dill==0.3.9
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker==7.1.0
dpath==2.2.0
dulwich==0.22.3
dvc==3.55.2
dvc-data==3.16.6
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
entrypoints==0.4
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
flatten-dict==0.4.2
flufl.lock==8.1.0
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
funcy==2.0
fusepy==3.0.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
google-pasta==0.2.0
gql==3.5.0
grandalf==0.8
graphene==3.3
graphql-core==3.2.5
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.67.0
gto==1.7.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
identify==2.6.1
idna==3.10
importlib_metadata==8.4.0
isort==5.13.2
iterative-telemetry==0.0.9
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.6.0
kiwisolver==1.4.7
kombu==5.4.2
libclang==18.1.1
loguru==0.7.2
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.1
marshmallow==3.22.0
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mlflow==2.17.0
mlflow-skinny==2.17.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.4.1
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
omegaconf==2.3.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opt_einsum==3.4.0
optree==0.13.0
orjson==3.10.7
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pathvalidate==3.2.1
pillow==11.0.0
platformdirs==3.11.0
pre_commit==4.0.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
pyahocorasick==2.1.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydot==3.0.2
pygit2==1.16.0
Pygments==2.18.0
pygtrie==2.5.0
pylint==3.3.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.3
scikit-learn==1.5.2
scipy==1.14.1
scmrepo==3.3.8
seaborn==0.13.2
semver==3.0.2
setuptools==75.2.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.1
sqltrie==0.11.1
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
termcolor==2.5.0
textsearch==0.0.24
tf_keras==2.16.0
threadpoolctl==3.5.0
tomlkit==0.13.2
torch==2.2.2
tqdm==4.66.5
treelib==1.7.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.26.6
voluptuous==0.15.2
wcwidth==0.2.13
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
yarl==1.15.3
zc.lockfile==3.0.post1
zipp==3.20.2
